--- src/Atlas.Application/Countries/DependencyInjection.cs
@@ -1,27 +1,0 @@
-// Copyright (c) KappaDuck. All rights reserved.
-// The source code is licensed under MIT License.
-
-using Atlas.Application.Countries.Commands;
-using Atlas.Application.Countries.Queries;
-using Microsoft.Extensions.DependencyInjection;
-using System.Diagnostics.CodeAnalysis;
-
-namespace Atlas.Application.Countries;
-
-[ExcludeFromCodeCoverage]
-internal static class DependencyInjection
-{
-    extension(IServiceCollection services)
-    {
-        internal void AddCountries()
-        {
-            services.AddSingleton<IGetAllCountries, GetAllCountries>();
-            services.AddSingleton<IGetCountry, GetCountry>();
-            services.AddSingleton<IGetDailyCountry, GetDailyCountry>();
-            services.AddSingleton<IGetDailyFlag, GetDailyFlag>();
-            services.AddSingleton<ILookupCountries, LookupCountries>();
-            services.AddSingleton<IRandomizeCountry, RandomizeCountry>();
-            services.AddSingleton<IGuessCountry, GuessCountry>();
-        }
-    }
-}
