+++ src/Atlas.Application/DependencyInjection.cs
@@ -1,8 +1,9 @@
 // Copyright (c) KappaDuck. All rights reserved.
 // The source code is licensed under MIT License.
 
-using Atlas.Application.Changelog;
-using Atlas.Application.Countries;
+using Atlas.Application.Changelog.Queries;
+using Atlas.Application.Countries.Commands;
+using Atlas.Application.Countries.Queries;
 using Microsoft.Extensions.DependencyInjection;
 using System.Diagnostics.CodeAnalysis;
 
@@ -17,6 +18,19 @@ public static class DependencyInjection
         {
             services.AddChangelog();
             services.AddCountries();
+        }
+
+        private void AddChangelog() => services.AddSingleton<IGetChangelog, GetChangelog>();
+
+        private void AddCountries()
+        {
+            services.AddSingleton<IGetAllCountries, GetAllCountries>();
+            services.AddSingleton<IGetCountry, GetCountry>();
+            services.AddSingleton<IGetDailyCountry, GetDailyCountry>();
+            services.AddSingleton<IGetDailyFlag, GetDailyFlag>();
+            services.AddSingleton<ILookupCountries, LookupCountries>();
+            services.AddSingleton<IRandomizeCountry, RandomizeCountry>();
+            services.AddSingleton<IGuessCountry, GuessCountry>();
         }
     }
 }
