+++ src/Web.App/Components/Modals/SettingsModal.razor
@@ -1,20 +1,20 @@
 ﻿@using Web.App.Settings.Components
 
 @inherits ModalBase
-@inject IStringLocalizer<Resources> localizer
+@inject IStringLocalizer<Resources> Localizer
 
 <CascadingValue Value="this" IsFixed="true">
     <dialog @ref="_dialog" class="modal">
         <div class="settings">
             <div class="header">
-                <h1>@localizer[Resources.Settings]</h1>
+                <h1>@Localizer[Resources.Settings]</h1>
                 <i class="@Icons.Times close" @onclick="() => Close(_dialog)"></i>
             </div>
             <div class="body">
                 <ul class="tabs" role="list">
                     @foreach ((TabItem tab, string name) in GetTabs())
                     {
-                        <li class="tab @IsActive(tab)" @onclick="() => SelectTab(tab)">@localizer[name]</li>
+                        <li class="tab @IsActive(tab)" @onclick="() => SelectTab(tab)">@Localizer[name]</li>
                     }
                 </ul>
                 <div class="content @GetTabCss()">
