+++ src/Web.App/Games/Countries/Components/CountryGuesses.razor.cs
@@ -23,5 +23,5 @@ public sealed partial class CountryGuesses
 
     private static string DirectionCss(bool isFound) => isFound ? $"found {Icons.Check}" : Icons.ArrowUp;
 
-    private static string IsSameContinentCss(bool IsSameContinent) => IsSameContinent ? $"found {Icons.Check}" : Icons.Times;
+    private static string IsSameContinentCss(bool isSameContinent) => isSameContinent ? $"found {Icons.Check}" : Icons.Times;
 }
