+++ src/Web.App/Index.razor
@@ -1,20 +1,20 @@
 @page "/"
 
-@inject IStringLocalizer<Resources> localizer
+@inject IStringLocalizer<Resources> Localizer
 
 <div class="container">
 	<div class="game-list">
-		<h2><i class="@Icons.Flag"></i><span>@localizer[Resources.Flags]</span></h2>
+		<h2><i class="@Icons.Flag"></i><span>@Localizer[Resources.Flags]</span></h2>
 		<div class="games">
-			<a class="btn-link" href="/flags/daily">@localizer[Resources.Daily]</a>
-			<a class="btn-link" href="/flags/random">@localizer[Resources.Random]</a>
+			<a class="btn-link" href="/flags/daily">@Localizer[Resources.Daily]</a>
+			<a class="btn-link" href="/flags/random">@Localizer[Resources.Random]</a>
 		</div>
 	</div>
 	<div class="game-list">
-		<h2><i class="@Icons.Earth"></i><span>@localizer[Resources.Countries]</span></h2>
+		<h2><i class="@Icons.Earth"></i><span>@Localizer[Resources.Countries]</span></h2>
 		<div class="games">
-			<a class="btn-link" href="/countries/daily">@localizer[Resources.Daily]</a>
-			<a class="btn-link" href="/countries/random">@localizer[Resources.Random]</a>
+			<a class="btn-link" href="/countries/daily">@Localizer[Resources.Daily]</a>
+			<a class="btn-link" href="/countries/random">@Localizer[Resources.Random]</a>
 		</div>
 	</div>
 </div>
