+++ src/Web.App/Settings/Components/GeneralSettings.razor
@@ -1,31 +1,31 @@
-﻿@inject IStringLocalizer<Resources> localizer
+﻿@inject IStringLocalizer<Resources> Localizer
 
-<h2>@localizer[Resources.Display]</h2>
+<h2>@Localizer[Resources.Display]</h2>
 <div class="display-settings">
     <div class="display-item">
         <div class="theme">
-            <label for="theme">@localizer[Resources.Theme]</label>
+            <label id="theme" for="theme">@Localizer[Resources.Theme]</label>
             <select @bind="AppSettings.Theme">
                 @foreach ((Theme theme, string name) in GetThemes())
                 {
-                    <option value="@theme">@localizer[name]</option>
+                    <option value="@theme">@Localizer[name]</option>
                 }
             </select>
         </div>
     </div>
     <div class="display-item">
         <div class="language">
-            <label for="language">@localizer[Resources.Language]</label>
+            <label id="language" for="language">@Localizer[Resources.Language]</label>
             <select @bind="AppSettings.Language">
                 @foreach ((Language language, string name) in GetLanguages())
                 {
-                    <option value="@language">@localizer[name]</option>
+                    <option value="@language">@Localizer[name]</option>
                 }
             </select>
         </div>
         <div class="alert info">
             <i class="@Icons.InfoCircle"></i>
-            <span>@localizer[Resources.ChangingLanguageWarning]</span>
+            <span>@Localizer[Resources.ChangingLanguageWarning]</span>
         </div>
     </div>
 </div>
