+++ src/Web.App/Settings/Components/InputDifficulty.razor
@@ -1,12 +1,12 @@
 ﻿@typeparam T where T : struct, Enum
 
-@inject IStringLocalizer<Resources> localizer
+@inject IStringLocalizer<Resources> Localizer
 
 <div class="radio-group">
     @foreach ((T difficulty, string name) in GetDifficulties())
     {
         <div class="radio">
-            <label for="@GetRadioId(name)">@localizer[name]</label>
+            <label for="@GetRadioId(name)">@Localizer[name]</label>
             <input id="@GetRadioId(name)"
                    name="@GroupName"
                    type="radio"
