+++ src/Web.App/wwwroot/scripts/settings.ts
@@ -1,5 +1,5 @@
 export function changeTheme(theme: string): void {
-  const app = document.querySelector('#app');
+  const app: Element | null = document.querySelector('#app');
 
   app?.setAttribute('data-theme', theme);
 }
